Added drag and drop support to creating sticks
Added a feature where a stick can be created by dragging the mouse between two balls, in addition to by clicking one to select it and then clicking the other. Fixed an issue where a ball could be connected to itself with a stick and where multiple sticks could be connected between the same balls.

Fixed access modifiers to be as restrictive as possible.

Adjusted stick physics iterations to be higher to combat the appearance of stretchiness from the approximation.